+++ Changes.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Encloses inner registry into headcode::logger namespace to avoid symbol conflict
 - Concurrent threading test.
 
 ## [1.0.0] - 2021-02-03